Moved existing Tuque code into a new location and Namespace Tuque\Fedora\v3.
Unit Tests work now for Tuque\Fedora\v3 code, as well as for the
RepositoryFactory object which is responsible for implementing the correct
fedora from the given config.

Work still needs to be done in converting the Objects in Tuque\Fedora\v3 to use
interfaces rather than abstract objects.

SimpleCache has been pulled out of the Tuque\Fedora\v3 code base since it is
generic enought to be used across Fedora 3 and Fedora 4.

For now many classes/interfaces are implemented in the Abstract*.php files in
the root directory; this is because creating files with similar names to the
ones that existed before will mess with gits history. They will be split out
later, once the code gets merged.